finishing game session
