+++ nbproject/private/private.xml
@@ -6,9 +6,9 @@
             <file>file:/C:/Users/zeyad_maamoun/Desktop/TicTacToeServer/src/network/ClientHandler.java</file>
         </group>
         <group>
-            <file>file:/C:/Users/mahmo/OneDrive/Documents/NetBeansProjects/TicTacToeServer/src/network/ClientHandler.java</file>
-            <file>file:/C:/Users/mahmo/OneDrive/Documents/NetBeansProjects/TicTacToeServer/src/alphaserver/FXMLDocumentController.java</file>
-            <file>file:/C:/Users/mahmo/OneDrive/Documents/NetBeansProjects/TicTacToeServer/src/alphaserver/AlphaServer.java</file>
+            <file>file:/C:/Users/mahmo/Downloads/netbeans/TicTacToeServer/src/network/ClientHandler.java</file>
+            <file>file:/C:/Users/mahmo/Downloads/netbeans/TicTacToeServer/src/network/GameSession.java</file>
+            <file>file:/C:/Users/mahmo/Downloads/netbeans/TicTacToeServer/src/alphaserver/AlphaServer.java</file>
         </group>
     </open-files>
 </project-private>
