+++ src/network/ClientHandler.java
@@ -196,7 +196,6 @@ public class ClientHandler extends Thread {
 
     }
 //accept 
-
     private void playerResponseHandler(JSONObject jsonMessage) {
         ClientHandler player1 = null;
         ClientHandler player2 = null;
@@ -227,7 +226,6 @@ public class ClientHandler extends Thread {
 
     }
 //reject
-
     private void playerResponsetHandler(JSONObject jsonMessage) {
         String toPlayer = jsonMessage.getString("toplayer");
         for (int i = 0; i < clients.size(); i++) {
@@ -240,182 +238,5 @@ public class ClientHandler extends Thread {
                 }
             }
         }
-    }
-}
-
-class GameSession extends Thread {
-
-    char currentPlayer = 'X';
-    boolean isPlayerOneTurn;
-    char[][] board = new char[3][3];
-    ClientHandler player1;
-    ClientHandler player2;
-    boolean isRunning = true;
-
-    public GameSession(ClientHandler player1, ClientHandler player2) {
-        this.player1 = player1;
-        this.player2 = player2;
-        initializeGame();
-        start();
-    }
-
-    @Override
-    public void run() {
-        try {
-            JSONObject obj;
-            // Initial notification to both players
-            JSONObject startObj = new JSONObject();
-            startObj.put("command", "start");
-            player1.mouth.writeUTF(startObj.toString());
-            while (isRunning) {
-                // Handle game messages from both players
-                if (player1.ear.available() > 0) {
-                    String move = player1.ear.readUTF();
-                    obj = new JSONObject(move);
-                    int col = obj.getInt("col");
-                    int row = obj.getInt("row");
-                    board[row][col] = currentPlayer;
-                    player2.mouth.writeUTF(move);
-                    if (checkWinner()) {
-                        if (currentPlayer == 'X') {
-                            notifyPlayersSomeoneWon(player1, player2);
-                            System.out.println("server says X won");
-                        } else {
-                            notifyPlayersSomeoneWon(player2, player1);
-                            System.out.println("server says O won");
-                            break;
-                        }
-                    } else if (isBoardFull()) {
-                        notifyPlayersDraw();
-                        System.out.println("server says board full");
-                        break;
-                    }
-                    isPlayerOneTurn = !isPlayerOneTurn;
-                    currentPlayer = (currentPlayer == 'X') ? 'O' : 'X';
-                    for (int i = 0; i < 3; i++) {
-                        for (int j = 0; j < 3; j++) {
-                            if(board[i][j]!=' '){
-                                System.out.println(board[i][j]);
-                            }else{
-                                System.out.println("-");
-                            }
-                        }
-                    }
-                    System.out.println("============");
-                }
-
-                if (player2.ear.available() > 0) {
-                    String move = player2.ear.readUTF();
-
-                    obj = new JSONObject(move);
-                    int col = obj.getInt("col");
-                    int row = obj.getInt("row");
-
-                    board[row][col] = currentPlayer;
-                    player1.mouth.writeUTF(move);
-
-                    if (checkWinner()) {
-                        if (currentPlayer == 'X') {
-                            notifyPlayersSomeoneWon(player1, player2);
-                            System.out.println("server says X won");
-                        } else {
-                            notifyPlayersSomeoneWon(player2, player1);
-                            System.out.println("server says O won");
-                            break;
-                        }
-                    } else if (isBoardFull()) {
-                        notifyPlayersDraw();
-                        System.out.println("server says board full");
-                        break;
-                    }
-                    isPlayerOneTurn = !isPlayerOneTurn;
-                    currentPlayer = (currentPlayer == 'X') ? 'O' : 'X';
-                    for (int i = 0; i < 3; i++) {
-                        for (int j = 0; j < 3; j++) {
-                            if(board[i][j]!=' '){
-                                System.out.println(board[i][j]);
-                            }else{
-                                System.out.println("-");
-                            }
-                        }
-                    }
-                    System.out.println("============");
-                }
-
-                Thread.sleep(50); // Small delay to prevent CPU overuse
-            }
-
-        } catch (IOException | InterruptedException ex) {
-            Logger.getLogger(GameSession.class.getName()).log(Level.SEVERE, null, ex);
-        } finally {
-            // Reset game state when session ends
-
-            isRunning = false;
-        }
-    }
-
-    public final void initializeGame() {
-        for (int i = 0; i < 3; i++) {
-            for (int j = 0; j < 3; j++) {
-                board[i][j] = ' ';
-            }
-        }
-    }
-
-    public boolean checkWinner() {
-        // Check rows and columns
-        for (int i = 0; i < 3; i++) {
-            if (board[i][0] == currentPlayer && board[i][1] == currentPlayer && board[i][2] == currentPlayer) {
-                return true;
-            }
-            if (board[0][i] == currentPlayer && board[1][i] == currentPlayer && board[2][i] == currentPlayer) {
-                return true;
-            }
-        }
-        // after checking rows and columns we check the diagonal.
-        if (board[0][0] == currentPlayer && board[1][1] == currentPlayer && board[2][2] == currentPlayer) {
-            return true;
-        }
-        if (board[0][2] == currentPlayer && board[1][1] == currentPlayer && board[2][0] == currentPlayer) {
-            return true;
-        }
-        return false;
-    }
-
-    void notifyPlayersSomeoneWon(ClientHandler winner, ClientHandler loser) {
-        JSONObject obj = new JSONObject();
-        obj.put("command", "win");
-        obj.put("winner", winner.username);
-        obj.put("loser", loser.username);
-        try {
-            player1.mouth.writeUTF(obj.toString());
-            player2.mouth.writeUTF(obj.toString());
-        } catch (IOException ex) {
-            Logger.getLogger(GameSession.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-
-    void notifyPlayersDraw() {
-        JSONObject obj = new JSONObject();
-        obj.put("command", "draw");
-        obj.put("player1", player1.username);
-        obj.put("player2", player2.username);
-        try {
-            player1.mouth.writeUTF(obj.toString());
-            player2.mouth.writeUTF(obj.toString());
-        } catch (IOException ex) {
-            Logger.getLogger(GameSession.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-
-    public final boolean isBoardFull() {
-        for (int i = 0; i < 3; i++) {
-            for (int j = 0; j < 3; j++) {
-                if (board[i][j] == ' ') {
-                    return false;
-                }
-            }
-        }
-        return true;
     }
 }
